Stupid uncommented code

+++ src/application/Main.java
@@ -17,8 +17,8 @@ import javafx.scene.layout.BorderPane;
 public class Main extends Application {
 	//kich thuoc cua scene
 	    Scene scene;
-		public static final int width = 1000;
-		public static final int height = 600;
+		public static final double width = 1000;
+		public static final double height = 600;
 		 //gameLoop
 	    private double timeLast ;
 	    private int countFrame = 0;
